requirements.txt
Django==2.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.6
olefile==0.44
Pillow==4.3.0
pytz==2017.3
six==1.11.0
